Refactored CurrentSearch, Added SearchTermsList
