+++ src/client/search/Search.js
@@ -1,8 +1,5 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { addSearchTerm } from './SearchActions';
-
-let nextSearchTermId = 0;
+import Current from './Current';
 
 const Search = () => {
 	return(
@@ -13,62 +10,10 @@ const Search = () => {
 	);
 };
 
-let Current = ({ dispatch }) => {
-	let showSearchkeyword = false;
-	let searchKeyword;
-	let searchKeywordContainer;
-	return(
-		<div className="row ui raised segment">
-			<div style={{cursor: 'text'}} onClick={() => {
-				showSearchkeyword = !showSearchkeyword;
-				if(showSearchkeyword) {
-					$("#searchKeywordContainer").slideDown("fast", () => {
-						searchKeyword.focus();
-					});
-				}
-			}}>
-				<i className="icon search"></i>
-
-				<div id="searchKeywordContainer" className="ui fluid big transparent input" 
-					style={{
-						display:'none', 
-						paddingLeft:20,
-						paddingTop:10,
-						marginTop:10,
-						borderTop:1,
-						borderTop: '2px dashed #D3D5D8'
-				}}>
-					<input id="searchKeyword" type="text" placeholder="Search a keyword or hashtag" 
-						ref={(node) => {
-							searchKeyword = node;
-						}}
-						onBlur={() => {
-							$("#searchKeywordContainer").slideUp("fast", () => {
-								showSearchkeyword = false;
-							});
-						}}
-						onKeyDown={(e) => {
-							if(e.keyCode == 13) {
-								dispatch(addSearchTerm(nextSearchTermId++, searchKeyword.value));
-								searchKeyword.value = '';
-							}
-						}}/>
-					<i className="link remove circle icon"></i>
-				</div>
-			</div>
-			
-			{/* action buttons (ie save filter)*/}
-			<div>
-			</div>
-		</div>
-	);
-};
-Current = connect()(Current);
-
 const Filters = () => {
 	return(
 		<div className="row">
-			Content
+			Filters will go here
 		</div>
 	);
 };
